code fixes and moved and split e2e tests
