+++ gef-ui/server/controllers/automatic-cover/index.js
@@ -39,7 +39,7 @@ const validateAutomaticCover = async (req, res, next) => {
     const { user } = session;
     const application = await api.getApplication(dealId);
     const { eligibility } = application;
-    const { _id } = user;
+    const { _id: editorId } = user;
 
     const automaticCoverErrors = getValidationErrors(body, eligibility.criteria);
     const coverType = deriveCoverType(body, eligibility.criteria);
@@ -76,7 +76,7 @@ const validateAutomaticCover = async (req, res, next) => {
       eligibility: {
         criteria: newAnswers,
       },
-      editorId: _id,
+      editorId,
     };
 
     await api.updateApplication(dealId, applicationUpdate);
