+++ gef-ui/server/controllers/review-decision/index.js
@@ -7,7 +7,7 @@ const acceptUkefDecision = async (req, res) => {
   const { dealId } = req.params;
   const { decision } = req.body;
   const { user } = req.session;
-  const { _id } = user;
+  const { _id: editorId } = user;
 
   try {
     if (decision) {
@@ -15,7 +15,7 @@ const acceptUkefDecision = async (req, res) => {
       const facilities = await api.getFacilities(dealId);
 
       application.ukefDecisionAccepted = true;
-      application.editorId = _id;
+      application.editorId = editorId;
       await api.updateApplication(dealId, application);
       /**
        * If issued facilities available then redirect user to
